+++ problem/problem_common.go
@@ -1,6 +1,8 @@
 package problem
 
 import (
+	"github.com/roidaradal/fn/ds"
+	"github.com/roidaradal/fn/list"
 	"github.com/roidaradal/opt/data"
 	"github.com/roidaradal/opt/discrete"
 	"github.com/roidaradal/opt/fn"
@@ -25,8 +27,8 @@ func newBinProblem(name string) (*discrete.Problem, *data.Bins) {
 	return p, cfg
 }
 
-// Common steps for creating Graph Cover problems
-func newGraphCoverProblem(name string, variablesFn data.GraphVariablesFn) (*discrete.Problem, *data.Graph) {
+// Common steps for creating Graph Subset problems (Graph Cover, Independent Set)
+func newGraphSubsetProblem(name string, variablesFn data.GraphVariablesFn) (*discrete.Problem, *data.Graph) {
 	graph := data.NewUndirectedGraph(name)
 	if graph == nil {
 		return nil, nil
@@ -39,8 +41,44 @@ func newGraphCoverProblem(name string, variablesFn data.GraphVariablesFn) (*disc
 	p.Variables = discrete.Variables(variables)
 	p.AddVariableDomains(discrete.BooleanDomain())
 
-	p.Goal = discrete.Minimize
 	p.ObjectiveFn = fn.ScoreSubsetSize
 	p.SolutionStringFn = fn.StringSubset(variables)
+	return p, graph
+}
+
+// Common steps for creating Graph Cover problems
+func newGraphCoverProblem(name string, variablesFn data.GraphVariablesFn) (*discrete.Problem, *data.Graph) {
+	p, graph := newGraphSubsetProblem(name, variablesFn)
+	if p == nil || graph == nil {
+		return nil, nil
+	}
+
+	p.Goal = discrete.Minimize
+	return p, graph
+}
+
+// Common steps for creating Independent Set problem
+func newIndependentSetProblem(name string) (*discrete.Problem, *data.Graph) {
+	p, graph := newGraphSubsetProblem(name, data.GraphVertices)
+	if p == nil || graph == nil {
+		return nil, nil
+	}
+
+	p.AddUniversalConstraint(func(solution *discrete.Solution) bool {
+		// Check that subset of vertices forms an independent set:
+		// none of the vertices are connected to each other
+		vertices := list.MapList(fn.AsSubset(solution), graph.Vertices)
+		// IsIndependentSet check
+		vertexSet := ds.SetFrom(vertices)
+		for _, vertex := range vertices {
+			adjacent := ds.SetFrom(graph.Neighbors(vertex))
+			if vertexSet.Intersection(adjacent).NotEmpty() {
+				return false
+			}
+		}
+		return true
+	})
+
+	p.Goal = discrete.Maximize
 	return p, graph
 }